refactor: move habit code into features directory
