+++ lib/features/habits/ui/widgets/habit_category.dart
