+++ lib/features/habits/ui/widgets/habit_completion_card.dart
