+++ lib/features/habits/ui/widgets/habit_page_app_bar.dart
@@ -2,10 +2,10 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:lotti/blocs/habits/habits_cubit.dart';
 import 'package:lotti/blocs/habits/habits_state.dart';
+import 'package:lotti/features/habits/ui/widgets/habits_filter.dart';
+import 'package:lotti/features/habits/ui/widgets/status_segmented_control.dart';
 import 'package:lotti/themes/theme.dart';
 import 'package:lotti/widgets/charts/habits/habit_completion_rate_chart.dart';
-import 'package:lotti/widgets/habits/habits_filter.dart';
-import 'package:lotti/widgets/habits/status_segmented_control.dart';
 import 'package:lotti/widgets/settings/settings_icon.dart';
 import 'package:material_design_icons_flutter/material_design_icons_flutter.dart';
